backend-stock-data/backend/requirements.txt
# backend/requirements.txt
# ─── Core ─────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0

# ─── Database ─────────────────
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0          # PostgreSQL async driver
alembic==1.13.1          # migrations

# ─── Evernote ─────────────────
evernote3==1.25.14        # Evernote SDK for Python 3

# ─── Stock Data ───────────────
FinanceDataReader==0.9.66  # OHLCV (한국/미국/글로벌)
pykrx==1.0.45              # 거래대금, 시가총액, 투자자별 수급
TA-Lib==0.4.32             # 기술적 보조지표 130+종

# ─── Utilities ────────────────
python-dotenv==1.0.0
httpx==0.26.0             # async HTTP client
cryptography==42.0.0      # token encryption
numpy>=1.24.0             # TA-Lib 의존

# ─── Phase 2-3 (uncomment when needed) ─────
# dart-fss==0.4.5            # DART 재무제표/공시
# OpenDartReader==0.2.1      # DART 공시 조회
# telethon==1.34.0           # Telegram client
# beautifulsoup4==4.12.3     # web scraping
# anthropic==0.18.0          # Claude AI API
